// src/data/types/ProjectData.ts
////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

import { LogoKey } from "../keys/LogoKey";

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////

export type ProjectData = {
  title: string;
  imageAlt: string;
  imageSrc: string;
  videoSrc: string;
  summary: string;
  description: string;
  repositoryLinkHref: string;
  repositoryLinkText: string;
  logoKeyArray: LogoKey[];
}

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////